workvivo-SIEM-python/requirements.txt
requests>=2.28.0
python-dotenv>=0.19.0
pytz>=2022.1 